+++ src/main/java/com/prgrms/be/intermark/domain/seatgrade/service/SeatGradeService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.prgrms.be.intermark.domain.musical.dto.MusicalSeatGradeCreateRequestDto;
+import com.prgrms.be.intermark.domain.musical.dto.MusicalSeatGradeCreateRequestDTO;
 import com.prgrms.be.intermark.domain.musical.model.Musical;
 import com.prgrms.be.intermark.domain.seatgrade.model.SeatGrade;
 import com.prgrms.be.intermark.domain.seatgrade.repository.SeatGradeRepository;
@@ -19,7 +19,7 @@ public class SeatGradeService {
 	private final SeatGradeRepository seatGradeRepository;
 
 	@Transactional
-	public void saveSeatGrade(List<MusicalSeatGradeCreateRequestDto> createRequestDTOs, Musical musical) {
+	public void saveSeatGrade(List<MusicalSeatGradeCreateRequestDTO> createRequestDTOs, Musical musical) {
 		createRequestDTOs
 			.forEach(seatGrade -> {
 				SeatGrade createdSeatGrade = seatGrade.toEntity();
